requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
